Fix test script import paths after relocation
