+++ test-scripts/README.md
@@ -1,2 +1,18 @@
-Put all test-*.mjs files in here 
-also track the existance and function of these files here
+# Test scripts
+
+All `test-*.mjs` scripts are stored in this folder.
+
+## Current scripts
+
+- `test-hash-restore.mjs` — hash restore test workflow.
+- `test-inspect-ri.mjs` — router info inspection helper/test.
+- `test-lease-set.mjs` — lease set parsing/validation test.
+- `test-notbob-lookup.mjs` — notbob lookup test.
+- `test-ntcp2-connect.mjs` — NTCP2 connection test.
+- `test-ntcp2-loopback.mjs` — NTCP2 loopback test.
+- `test-parse-ri.mjs` — router info parsing test.
+- `test-peer-discovery.mjs` — peer discovery behavior test.
+- `test-ri-crypto.mjs` — router info cryptographic checks.
+- `test-ri-verify.mjs` — router info verification test.
+- `test-siphash.mjs` — siphash test.
+- `test-with-logs.mjs` — test run with extended logs.
