+++ test-scripts/test-inspect-ri.mjs
