+++ test-scripts/test-peer-discovery.mjs
