+++ test-scripts/test-with-logs.mjs
